+++ src/main/java/com/insightml/models/trees/SplitFinderContext.java
@@ -25,7 +25,7 @@ public final class SplitFinderContext {
 	public final double[] expected;
 	public final double[] weights;
 	public final String[] featureNames;
-	public final double[][] features;
+	public final float[][] features;
 	public final int[][] orderedInstances;
 
 	@Nullable
